Add files via upload
Rock vs. Mine Prediction.
During wars there are many mines distributed in the ocean by the enemies to harm friendly ships and submarines. It is important to detect the positions of these mines to navigate their way safely around the mines. All submarines have SONAR(Sound Navigation and Ranging) systems which are used to detect objects. But the problem is that there are many rocks among the mines and this causes hindrance while detecting them. We have a dataset which has many internal feature metrics and a label telling us if the object is a rock or a mine. We need to predict the occurence of mines using these features.