Add files via upload
We have a file containing patient records, which tells us whether a patient is diabetic or not. The feature variables are - Pregnancies, Glucose, Blood Pressure, Skin thickness, Insulin, BMI, Diabetes Pedigree Function and Age. Outcome is the target variable with a value of 0 and 1 for Non-Diabetic and Diabetic patients respectively.